requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.csv==1.0.7
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.0
django-celery==3.3.0
django-cors-headers==2.5.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-filter==2.1.0
django-formtools==2.1
django-guardian==1.5.1
django-import-export==1.2.0
django-redis==4.10.0
django-reversion==3.0.3
djangorestframework==3.9.3
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
et-xmlfile==1.0.1
future==0.17.1
httpie==1.0.2
httplib2==0.12.3
idna==2.8
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
kombu==3.0.37
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
Pygments==2.4.2
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
qiniu==7.2.4
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
sentry-sdk==0.7.10
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.24.3
uWSGI==2.0.18
vine==1.3.0
xlrd==1.2.0
XlsxWriter==1.1.7
xlwt==1.3.0
